+++ Task19UI/src/app/shared/modals/registry-modal/registry.modal.ts
@@ -1,5 +1,6 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Input, OnInit, Output} from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
+import { UserModel } from '../../models';
 
 @Component({
     selector: 'app-registry-modal',
