+++ Task19UI/src/app/shared/services/group.services.ts
@@ -7,7 +7,7 @@ import { HttpClient } from '@angular/common/http';
 @Injectable({
   providedIn: 'root'
 })
-export class groupService {
+export class GroupService {
   errorMessage: String = "HttpError";
  private url = "/group/";
